bug fix for filling up int

+++ common/smart_loop_logger.hh
@@ -1,7 +1,7 @@
 #include <cmath>
 #include <spdlog/spdlog.h>
 
-void smart_loop_logger(int nTotal, int iLoop, int each_display_percent=10){
+void smart_loop_logger(long nTotal, long iLoop, int each_display_percent=10){
     if (
            ((100 * iLoop / nTotal % each_display_percent == 0) && ((100 * iLoop / nTotal) != (100 * (iLoop-1) / nTotal))) ||
            ((int(std::log10(iLoop-1))) != (int(std::log10(iLoop))))
